Backend: correctness and robustness fixes from full audit
- Part partial-update no longer nulls out unit/reorderLevel (fixes a 500
  from NPE + NOT NULL on PATCH /parts/{id} with a subset of fields).
- Appointment update now persists notes (was silently dropped).
- Map IllegalArgumentException -> 400 and AccessDeniedException -> 403 so
  domain guards and @PreAuthorize denials no longer surface as 500.
- Reject negative unitCost via @DecimalMin on stock-in and consume-part.
- Enforce "assigned technician only" on appointment confirm/start/
  complete/cancel (mirrors the service-request rule).
- Error response timestamp is now epoch millis, not nanosecond-of-second.

+++ backend/src/main/java/com/fixonaut/backend/common/api/ApiErrorResponse.java
@@ -3,7 +3,7 @@ package com.fixonaut.backend.common.api;
 import java.util.Map;
 
 public record ApiErrorResponse(
-        Integer timestamp,
+        long timestamp,
         int status,
         String code,
         String message,
